+++ src/config.c
@@ -107,7 +107,7 @@ TbBool parameter_is_number(const char* parstr) {
     return true;
 }
 
-int get_conf_line(const char *buf, long *pos, long buflen, char *dst, long dstlen)
+int get_conf_line(const char *buf, int32_t *pos, long buflen, char *dst, long dstlen)
 {
     SYNCDBG(19,"Starting");
     if ((*pos) >= buflen) return ccr_endOfFile;
@@ -143,7 +143,7 @@ int get_conf_line(const char *buf, long *pos, long buflen, char *dst, long dstle
 }
 
 
-TbBool skip_conf_to_next_line(const char *buf,long *pos,long buflen)
+TbBool skip_conf_to_next_line(const char *buf,int32_t *pos,long buflen)
 {
   // Skip to end of the line
   while ((*pos) < buflen)
@@ -162,7 +162,7 @@ TbBool skip_conf_to_next_line(const char *buf,long *pos,long buflen)
   return ((*pos) < buflen);
 }
 
-TbBool skip_conf_spaces(const char *buf, long *pos, long buflen)
+TbBool skip_conf_spaces(const char *buf, int32_t *pos, long buflen)
 {
   while ((*pos) < buflen)
   {
@@ -177,7 +177,7 @@ TbBool skip_conf_spaces(const char *buf, long *pos, long buflen)
  * Starts at position given with pos, and sets it to position of block data.
  * @return Returns 1 if the block is found, -1 if buffer exceeded.
  */
-short find_conf_block(const char *buf,long *pos,long buflen,const char *blockname)
+short find_conf_block(const char *buf,int32_t *pos,long buflen,const char *blockname)
 {
   text_line_number = 1;
   int blname_len = strlen(blockname);
@@ -223,7 +223,7 @@ short find_conf_block(const char *buf,long *pos,long buflen,const char *blocknam
  * Sets name and namelen to the block name and name length respectively.
  * Returns true on success, false when the block name is zero.
  */
-TbBool conf_get_block_name(const char * buf, long * pos, long buflen, const char ** name, int * namelen)
+TbBool conf_get_block_name(const char * buf, int32_t * pos, long buflen, const char ** name, int * namelen)
 {
   const long start = *pos;
   *name = NULL;
@@ -254,7 +254,7 @@ TbBool conf_get_block_name(const char * buf, long * pos, long buflen, const char
  * Sets name and namelen to the block name and name length respectively.
  * Returns true on success, false when no more blocks are found.
  */
-TbBool iterate_conf_blocks(const char * buf, long * pos, long buflen, const char ** name, int * namelen)
+TbBool iterate_conf_blocks(const char * buf, int32_t * pos, long buflen, const char ** name, int * namelen)
 {
   text_line_number = 1;
   *name = NULL;
@@ -306,7 +306,7 @@ TbBool iterate_conf_blocks(const char * buf, long * pos, long buflen, const char
  * If ccr_unrecognised is returned, that means the command wasn't recognized.
  * If ccr_endOfBlock   is returned, that means we've reached end of the INI block.
  */
-int recognize_conf_command(const char *buf,long *pos,long buflen,const struct NamedCommand commands[])
+int recognize_conf_command(const char *buf,int32_t *pos,long buflen,const struct NamedCommand commands[])
 {
     SYNCDBG(19,"Starting");
     if ((*pos) >= buflen) return ccr_endOfFile;
@@ -415,13 +415,13 @@ static int64_t get_datatype_max(uchar type)
         case dt_uint:
             return UINT_MAX;
         case dt_long:
-            return LONG_MAX;
+            return INT32_MAX;
         case dt_ulong:
-            return ULONG_MAX;
+            return UINT32_MAX;
         case dt_longlong:
-            return LLONG_MAX;
+            return INT64_MAX;
         case dt_ulonglong:
-            return ULLONG_MAX;
+            return UINT64_MAX;
         default:
             break;
     }
@@ -487,7 +487,7 @@ int64_t value_longflagsfield(const struct NamedField* named_field, const char* v
         return 0;
     }
 
-    long pos = 0;
+    int32_t pos = 0;
     long len = strlen(value_text);
     int i = 0;
     while (get_conf_parameter_single(value_text,&pos,len,word_buf,sizeof(word_buf)) > 0)
@@ -528,7 +528,7 @@ int64_t value_flagsfield(const struct NamedField* named_field, const char* value
         return 0;
     }
 
-    long pos = 0;
+    int32_t pos = 0;
     long len = strlen(value_text);
     int i = 0;
     while (get_conf_parameter_single(value_text,&pos,len,word_buf,sizeof(word_buf)) > 0)
@@ -755,25 +755,25 @@ void assign_default(const struct NamedField* named_field, int64_t value, const s
             *(unsigned int*)field = (unsigned int)value;
         break;
     case dt_long:
-        if (value < LONG_MIN || value > LONG_MAX)
+        if (value < LONG_MIN || value > INT32_MAX)
             NAMFIELDWRNLOG("Value out of range for signed long: %lld", value);
         else
             *(signed long*)field = (signed long)value;
         break;
     case dt_ulong:
-        if (value < 0 || value > ULONG_MAX)
+        if (value < 0 || value > UINT32_MAX)
             NAMFIELDWRNLOG("Value out of range for unsigned long: %lld", value);
         else
             *(unsigned long*)field = (unsigned long)value;
         break;
     case dt_longlong:
-        if (value < LLONG_MIN || value > LLONG_MAX)
+        if (value < LLONG_MIN || value > INT64_MAX)
             NAMFIELDWRNLOG("Value out of range for signed long long: %lld", value);
         else
             *(signed long long*)field = (signed long long)value;
         break;
     case dt_ulonglong:
-        if (value < 0 || value > ULLONG_MAX)
+        if (value < 0 || value > UINT64_MAX)
             NAMFIELDWRNLOG("Value out of range for unsigned long long: %lld", value);
         else
             *(unsigned long long*)field = (unsigned long long)value;
@@ -823,7 +823,7 @@ void assign_named_field_value(const struct NamedField* named_field, int64_t valu
  * If ccr_error        is returned, that means something went wrong.
  */
 
-int assign_conf_command_field(const char *buf,long *pos,long buflen,const struct NamedField commands[], const struct NamedFieldSet* named_fields_set, int idx, unsigned short flags, const char *config_textname)
+int assign_conf_command_field(const char *buf,int32_t *pos,long buflen,const struct NamedField commands[], const struct NamedFieldSet* named_fields_set, int idx, unsigned short flags, const char *config_textname)
 {
     SYNCDBG(19,"Starting");
     if ((*pos) >= buflen) return -1;
@@ -946,7 +946,7 @@ int assign_conf_command_field(const char *buf,long *pos,long buflen,const struct
 TbBool parse_named_field_block(const char *buf, long len, const char *config_textname, unsigned short flags,const char* blockname,
                          const struct NamedField named_field[], const struct NamedFieldSet* named_fields_set, int idx)
 {
-    long pos = 0;
+    int32_t pos = 0;
     int k = find_conf_block(buf, &pos, len, blockname);
     if (k < 0)
     {
@@ -1015,7 +1015,7 @@ void set_defaults(const struct NamedFieldSet* named_fields_set, const char *conf
 TbBool parse_named_field_blocks(char *buf, long len, const char *config_textname, unsigned short flags,
                                const struct NamedFieldSet* named_fields_set)
 {
-    long pos = 0;
+    int32_t pos = 0;
     // Initialize the array
     if ((flags & CnfLd_AcceptPartial) == 0)
     {
@@ -1049,7 +1049,7 @@ TbBool parse_named_field_blocks(char *buf, long len, const char *config_textname
     return true;
 }
 
-int get_conf_parameter_whole(const char *buf,long *pos,long buflen,char *dst,long dstlen)
+int get_conf_parameter_whole(const char *buf,int32_t *pos,long buflen,char *dst,long dstlen)
 {
   int i;
   if ((*pos) >= buflen) return 0;
@@ -1071,7 +1071,7 @@ int get_conf_parameter_whole(const char *buf,long *pos,long buflen,char *dst,lon
   return i;
 }
 
-int get_conf_parameter_single(const char *buf,long *pos,long buflen,char *dst,long dstlen)
+int get_conf_parameter_single(const char *buf,int32_t *pos,long buflen,char *dst,long dstlen)
 {
     int i;
     if ((*pos) >= buflen) return 0;
@@ -1100,7 +1100,7 @@ int get_conf_parameter_single(const char *buf,long *pos,long buflen,char *dst,lo
 /**
  * Returns parameter num from given NamedCommand array, or 0 if not found.
  */
-int recognize_conf_parameter(const char *buf,long *pos,long buflen,const struct NamedCommand commands[])
+int recognize_conf_parameter(const char *buf,int32_t *pos,long buflen,const struct NamedCommand commands[])
 {
   if ((*pos) >= buflen) return 0;
   // Skipping spaces after previous parameter
@@ -1419,7 +1419,7 @@ short get_level_fgroup(LevelNumber lvnum)
  * on success, returns a buffer which should be freed after use,
  * and sets ldsize into its size.
  */
-unsigned char *load_data_file_to_buffer(long *ldsize, short fgroup, const char *fmt_str, ...)
+unsigned char *load_data_file_to_buffer(int32_t *ldsize, short fgroup, const char *fmt_str, ...)
 {
   // Prepare file name
   va_list arg;
@@ -1586,7 +1586,7 @@ TbBool parse_credits_block(struct CreditsItem *credits,char *buf,char *buffer_en
   const char * block_name = "credits";
   // Find the block
   long len = buffer_end_pointer - buf;
-  long pos = 0;
+  int32_t pos = 0;
   int k = find_conf_block(buf, &pos, len, block_name);
   if (k < 0)
   {
